+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## FUTURE
+
+* [FIXED] - async listeners should always be called with the Emitter object as context (this keyword)
+
 ## v1.0.1
 
 * [FIXED] - use `promise.then` rather than `promise.return` method